Removed withAuthentication HOC, moved into PrivateRoute
