+++ 01-Login/src/components/PrivateRoute.js
@@ -1,14 +1,24 @@
-import React from "react";
+import React, { useEffect } from "react";
 import PropTypes from "prop-types";
 import { Route } from "react-router-dom";
-
-import withAuthentication from "../hocs/withAuthentication";
+import { useAuth0 } from "../react-auth0-spa";
 
 const PrivateRoute = ({ component: Component, path, ...rest }) => {
-  const ComponentWithAuthentication = withAuthentication(Component);
-  const render = props => (
-    <ComponentWithAuthentication path={path} {...props} />
-  );
+  const { isAuthenticated, loginWithRedirect } = useAuth0();
+
+  useEffect(() => {
+    const fn = async () => {
+      if (!isAuthenticated) {
+        await loginWithRedirect({
+          redirect_uri: window.location.origin,
+          appState: { targetUrl: path }
+        });
+      }
+    };
+    fn();
+  }, [isAuthenticated, loginWithRedirect, path]);
+
+  const render = props => <Component {...props} />;
 
   return <Route path={path} render={render} {...rest} />;
 };
